Remove 'Users' model from Django
We removed the Users model from Django because we decided all auth will go through the Django app for access to the OpenEMR Database.  Users of the mobile app will need to be configured through the admin portal hosted by Django and not OpenEMR. It is an important design decision that these be kept separate.